ADD: more tests
